deprecate test list cmd
